Migrate tests to Swift Testing
